Added launches error handling along with some minor refactoring

+++ SpaceX/Storage/Cache.swift
@@ -20,21 +20,25 @@ class Cache<V> where V: Codable {
   }
 
   func save(_ value: V) {
+    guard let fileURL = fileURL else { return }
     let data = try? JSONEncoder().encode(value)
     try? data?.write(to: fileURL)
   }
 
   var retrieve: V? {
+    guard let fileURL = fileURL else { return nil }
     guard FileManager.default.fileExists(atPath: fileURL.path),
           let data = try? Data(contentsOf: fileURL) else { return nil }
     return try? JSONDecoder().decode(V.self, from: data)
   }
 
   func clear() {
+    guard let folderURL = folderURL else { return }
     try? FileManager.default.removeItem(at: folderURL)
   }
 
   private func createDirectory() {
+    guard let folderURL = folderURL else { return }
     var isDirectory: ObjCBool = false
     if !FileManager.default.fileExists(atPath: folderURL.path,
                                        isDirectory: &isDirectory), !isDirectory.boolValue {
@@ -43,14 +47,14 @@ class Cache<V> where V: Codable {
     }
   }
 
-  private var fileURL: URL {
-    return folderURL.appendingPathComponent(filename)
+  private var fileURL: URL? {
+    return folderURL?.appendingPathComponent(filename)
   }
 
-  private var folderURL: URL {
-    let path = NSSearchPathForDirectoriesInDomains(.cachesDirectory,
+  private var folderURL: URL? {
+    guard let path = NSSearchPathForDirectoriesInDomains(.cachesDirectory,
                                                    .userDomainMask,
-                                                   true).first!
+                                                   true).first else { return nil }
     return URL(fileURLWithPath: path).appendingPathComponent(directory)
   }
